+++ src/main/java/com/cabinet/model/Prescription.java
@@ -1,8 +1,8 @@
 package com.cabinet.model;
 
 public class Prescription {
-    private int id;
-    private long  consultationId;
+    private Long id;           // Changed to Long object
+    private Long consultationId;  // Changed to Long object
     private String medicamentNom;
     private String posologie;
 
@@ -13,12 +13,15 @@ public class Prescription {
         this.posologie = posologie;
     }
 
-    public int getId() { return id; }
-    public void setId(int id) { this.id = id; }
-    public long getConsultationId() { return consultationId; }
-    public void setConsultationId(long consultationId) { this.consultationId = consultationId; }
+    public Long getId() { return id; }
+    public void setId(Long id) { this.id = id; }
+
+    public Long getConsultationId() { return consultationId; }
+    public void setConsultationId(Long consultationId) { this.consultationId = consultationId; }
+
     public String getMedicamentNom() { return medicamentNom; }
     public void setMedicamentNom(String medicamentNom) { this.medicamentNom = medicamentNom; }
+
     public String getPosologie() { return posologie; }
     public void setPosologie(String posologie) { this.posologie = posologie; }
 }
